# openapi.yaml
openapi: 3.0.3
info:
  title: API ИИ-гида музея Фаберже
  version: 0.1.0-mvp
  description: |
    Backend-API мобильного web-приложения (PWA) «ИИ-гид музея Фаберже».

    Посетитель сканирует **QR-код** на входе экспозиции, попадает в приложение
    без регистрации и установки, перемещается по **интерактивной карте**
    (зал → витрина → экспонат), а главное — может **сфотографировать экспонат**,
    получить о нём рассказ от ИИ-гида и **прослушать** его озвучку.

    Этот документ описывает контракт backend-сервиса (FastAPI на Yandex Cloud),
    к которому обращается фронтенд. Спецификация написана по дорожной карте MVP
    (25 мая – 25 июня) и служит единым контрактом между Frontend, Backend и ML.

    ## Ключевой пользовательский сценарий (бесшовное распознавание)

    ```
    Клик «Распознать экспонат»
      → камера (WebRTC)
      → снимок
      → POST /recognition (фото → YOLO → label_slug)
      → POST /guide/story (label_slug → raw_history из БД → YandexGPT → рассказ)
      → показ рассказа + 3–4 всплывающих вопроса
      → POST /speech («Прослушать», SpeechKit → аудио)
      → блок «Другие экспонаты зала» (GET /exhibits/{id}/related)
    ```

    ## Доменная модель

    Иерархия каталога повторяет физическую структуру музея:

    | Уровень   | Сущность   | Таблица БД   | Ключевые поля                                   |
    |-----------|------------|--------------|-------------------------------------------------|
...

...
    (через публичный API не отдаётся).

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security: []
paths:
...
  /halls:
    get:
      tags: [Карта и навигация]
      summary: Список залов
      description: Постраничный список залов музея для карты и навигации.
      operationId: listHalls
      security: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Список залов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HallListResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /halls/{hall_id}:
    get:
      tags: [Карта и навигация]
      summary: Получить зал
      description: >-
        Карточка зала с описанием и кнопкой «Спросить AI-гида» на фронте.
        Описание зала используется на уровне зала интерактивной карты.
      operationId: getHall
      security: []
      parameters:
        - $ref: '#/components/parameters/HallId'
      responses:
        '200':
          description: Данные зала.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HallDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /halls/{hall_id}/showcases:
    get:
      tags: [Карта и навигация]
      summary: Витрины зала
      description: Список витрин внутри указанного зала.
      operationId: listHallShowcases
      security: []
      parameters:
        - $ref: '#/components/parameters/HallId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Список витрин зала.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowcaseListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /halls/{hall_id}/exhibits:
    get:
      tags: [Карта и навигация]
      summary: Экспонаты зала
      description: >-
        Все экспонаты зала (по всем его витринам). Эта же выборка лежит в
        основе блока «Другие экспонаты зала» — соседние записи по `hall_id`.
      operationId: listHallExhibits
      security: []
      parameters:
        - $ref: '#/components/parameters/HallId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Список экспонатов зала.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /showcases/{showcase_id}:
    get:
      tags: [Карта и навигация]
      summary: Получить витрину
      operationId: getShowcase
      security: []
      parameters:
        - $ref: '#/components/parameters/ShowcaseId'
      responses:
        '200':
          description: Данные витрины.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowcaseDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /showcases/{showcase_id}/exhibits:
    get:
      tags: [Карта и навигация]
      summary: Экспонаты витрины
      operationId: listShowcaseExhibits
      security: []
      parameters:
        - $ref: '#/components/parameters/ShowcaseId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Список экспонатов витрины.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  # ─────────────────────────── Экспонаты ─────────────────────────────────
...
  /guide/story:
    post:
      tags: [ИИ-гид]
      summary: Сгенерировать рассказ об экспонате
      description: |
        Cloud Function берёт `raw_history` экспоната из БД и формирует промпт
        для **YandexGPT**:

        > «Напиши интересную историю для посетителя музея, используя данные:
        > `{raw_history}`, стиль: `{style}`, год: `{year_created}`.»

        Возвращает связный рассказ и **3–4 контекстных вопроса-подсказки**
        (всплывающие подсказки внизу чата для продолжения диалога). По флагу
        `include_audio` сразу прикладывает ссылку на озвучку (SpeechKit).

        Укажите **`exhibit_id` или `label_slug`** (хотя бы одно). `label_slug`
        удобен сразу после распознавания.
      operationId: generateStory
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoryRequest'
            examples:
              afterRecognition:
                summary: После распознавания (по label_slug)
                value:
                  label_slug: faberge_egg_coronation
                  style: engaging
                  language: ru
                  include_audio: false
                  max_questions: 4
      responses:
        '200':
          description: Сгенерированный рассказ и подсказки.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoryResponse'
              examples:
                story:
                  summary: Рассказ о Коронационном яйце
                  value:
                    exhibit_id: 101
                    label_slug: faberge_egg_coronation
                    style: engaging
                    text: >-
                      В 1897 году Карл Фаберже исполнил для Николая II особый
                      пасхальный подарок императрице Александре Фёдоровне…
                    suggested_questions:
                      - Сколько времени ушло на создание?
                      - Кому подарили это яйцо?
                      - Что за сюрприз спрятан внутри?
                      - Какие материалы использовал мастер?
                    audio_url: null
                    model: yandexgpt/latest
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '502':
          $ref: '#/components/responses/UpstreamError'
  /guide/chat:
    post:
      tags: [ИИ-гид]
      summary: Диалог с ИИ-гидом
      description: |
        Многоходовой диалог с гидом. Контекст задаётся экспонатом
        (`exhibit_id`/`label_slug`) либо залом (`hall_id` — для кнопки
        «Спросить AI-гида» на уровне зала). Часто `message` — это один из
        ранее предложенных `suggested_questions`.

        Состояние диалога хранится по `session_id`: если он не передан, сервер
        создаёт новый и возвращает его. Альтернативно можно передавать
        историю сообщений в `history`. Ответ снова содержит набор
        контекстных вопросов-подсказок.
      operationId: chatWithGuide
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            examples:
              followUp:
                summary: Уточняющий вопрос по экспонату
                value:
                  session_id: 2f6a9b3c-1d4e-4f8a-9c2b-5e7a1d0c3b66
                  context:
                    exhibit_id: 101
                  message: Кому подарили это яйцо?
      responses:
        '200':
          description: Ответ гида и новые подсказки.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
              examples:
                answer:
                  summary: Ответ на уточняющий вопрос
                  value:
                    session_id: 2f6a9b3c-1d4e-4f8a-9c2b-5e7a1d0c3b66
                    answer: >-
                      Это яйцо Николай II подарил своей супруге, императрице
                      Александре Фёдоровне, на Пасху 1897 года…
                    suggested_questions:
                      - А что подарили его матери?
                      - Где оно хранилось после революции?
                      - Сколько таких яиц всего создал Фаберже?
                    context:
                      exhibit_id: 101
                      label_slug: faberge_egg_coronation
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '502':
          $ref: '#/components/responses/UpstreamError'

  # ──────────────────────────── Озвучивание ──────────────────────────────
...
  /telemetry/events:
    post:
      tags: [Телеметрия]
      summary: '[Вне MVP] Отправить события'
      description: >-
        Пакетный приём анонимных пользовательских событий (просмотр экспоната,
        распознавание, сообщение в чате, прослушивание) — источник данных для
        административной аналитики.
      operationId: ingestEvents
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventBatch'
      responses:
        '202':
          description: События приняты в обработку.
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer-токен администратора. Только для эндпоинтов `/admin/**`.
  parameters:
    Limit:
      name: limit
      in: query
      description: Максимум элементов на странице.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      description: Смещение от начала выборки (постраничность).
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    HallId:
      name: hall_id
      in: path
      required: true
      description: Идентификатор зала.
      schema:
        type: integer
        minimum: 1
      example: 5
    ShowcaseId:
      name: showcase_id
      in: path
      required: true
      description: Идентификатор витрины.
      schema:
        type: integer
        minimum: 1
      example: 12
...
  responses:
    NotFound:
      description: Ресурс не найден.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Экспонат не найден.
...
    UnprocessableEntity:
      description: Ошибка валидации запроса (формат FastAPI).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTPValidationError'
    UpstreamError:
      description: >-
        Ошибка внешнего сервиса (YOLO / YandexGPT / SpeechKit).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Сервис распознавания временно недоступен.
  schemas:
...
    Hall:
      type: object
      required: [id, hall_number]
      properties:
        id:
          type: integer
          example: 5
        hall_number:
          type: integer
          description: Номер зала на плане экспозиции (уникальный).
          example: 5
        name:
          type: string
          nullable: true
          example: Синяя гостиная
        description:
          type: string
          nullable: true
          description: >-
            Описание зала для уровня зала на карте и для контекста ИИ-гида.
            Наполняется из плана экспозиции и путеводителя.
          example: >-
            Парадная гостиная Шуваловского дворца, где представлены девять
            императорских пасхальных яиц Фаберже.
        level:
          type: integer
          nullable: true
          description: Этаж / уровень в здании музея.
          example: 2
        cover_image_url:
          type: string
          format: uri
          nullable: true
          example: https://cdn.example.cloud/halls/blue/cover.jpg
        showcase_count:
          type: integer
          nullable: true
          example: 6
        exhibit_count:
          type: integer
          nullable: true
          example: 24
    HallBrief:
      type: object
      description: Краткая ссылка на зал (для вложений в карточку экспоната).
      required: [id, hall_number]
      properties:
        id:
          type: integer
          example: 5
        hall_number:
          type: integer
          example: 5
        name:
          type: string
          nullable: true
          example: Синяя гостиная
    HallDetail:
      allOf:
        - $ref: '#/components/schemas/Hall'
        - type: object
          properties:
            showcases:
              type: array
              description: Витрины зала (могут включаться по необходимости).
              items:
                $ref: '#/components/schemas/Showcase'
...
    HallListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Hall'
        total:
          type: integer
          example: 12
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0

    # ── Витрины ──
    Showcase:
      type: object
      required: [id, hall_id, showcase_number]
      properties:
        id:
          type: integer
          example: 12
        hall_id:
          type: integer
          example: 5
        showcase_number:
          type: integer
          description: Номер витрины внутри зала (уникален в пределах зала).
          example: 3
        name:
          type: string
          nullable: true
          example: Императорские пасхальные яйца
        exhibit_count:
          type: integer
          nullable: true
          example: 4
...
    ShowcaseDetail:
      allOf:
        - $ref: '#/components/schemas/Showcase'
        - type: object
          properties:
            hall:
              $ref: '#/components/schemas/HallBrief'
            exhibits:
              type: array
              items:
                $ref: '#/components/schemas/ExhibitSummary'
    ShowcaseListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Showcase'
        total:
          type: integer
          example: 6
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0

    # ── Экспонаты ──
...
    ExhibitSummary:
      type: object
      description: Краткое представление экспоната для списков и карусели.
      required: [id, name]
      properties:
        id:
          type: integer
          example: 101
        label_slug:
          type: string
          nullable: true
          example: faberge_egg_coronation
        name:
          type: string
          example: Яйцо «Коронационное»
        year_created:
          type: integer
          nullable: true
          example: 1897
        master_name:
          type: string
          nullable: true
          example: Михаил Перхин
        thumbnail_url:
          type: string
          format: uri
          nullable: true
          example: https://cdn.example.cloud/exhibits/coronation/thumb.jpg
        hall_id:
          type: integer
          nullable: true
          example: 5
        showcase_id:
          type: integer
          nullable: true
          example: 12
...
    ExhibitListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ExhibitSummary'
        total:
          type: integer
          example: 24
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0

    # ── Поиск ──
...
    GuideStyle:
      type: string
      description: Стиль генерируемого рассказа.
      enum: [engaging, historical, short, kids, expert]
      default: engaging
    GuideContext:
      type: object
      description: >-
        Контекст диалога. Укажите хотя бы одно из полей. `exhibit_id`/`label_slug`
        — контекст экспоната, `hall_id` — контекст зала.
      properties:
        exhibit_id:
          type: integer
          nullable: true
          example: 101
        label_slug:
          type: string
          nullable: true
          example: faberge_egg_coronation
        hall_id:
          type: integer
          nullable: true
          example: 5
    StoryRequest:
      type: object
      description: 'Запрос на генерацию рассказа. Укажите `exhibit_id` или `label_slug`.'
      properties:
        exhibit_id:
          type: integer
          nullable: true
          example: 101
        label_slug:
          type: string
          nullable: true
          example: faberge_egg_coronation
        style:
          $ref: '#/components/schemas/GuideStyle'
        language:
          type: string
          default: ru
          description: Язык рассказа (ISO-639-1).
          example: ru
        include_audio:
          type: boolean
          default: false
          description: Сразу синтезировать озвучку и вернуть `audio_url`.
        max_questions:
          type: integer
          minimum: 0
          maximum: 6
          default: 4
          description: Сколько вопросов-подсказок сгенерировать (по ТЗ — 3–4).
    StoryResponse:
      type: object
      required: [text, suggested_questions]
      properties:
        exhibit_id:
          type: integer
          nullable: true
          example: 101
        label_slug:
          type: string
          nullable: true
          example: faberge_egg_coronation
        style:
          $ref: '#/components/schemas/GuideStyle'
        text:
          type: string
          description: Сгенерированный YandexGPT рассказ.
        suggested_questions:
          type: array
          description: 3–4 контекстных вопроса-подсказки для продолжения диалога.
          items:
            type: string
          example:
            - Сколько времени ушло на создание?
            - Кому подарили это яйцо?
            - Что за сюрприз спрятан внутри?
        audio_url:
          type: string
          format: uri
          nullable: true
          description: Озвучка рассказа, если запрошен `include_audio`.
        model:
          type: string
          nullable: true
          example: yandexgpt/latest
        generated_at:
          type: string
          format: date-time
          nullable: true
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant, system]
          example: user
        content:
          type: string
          example: Кому подарили это яйцо?
    ChatRequest:
      type: object
      required: [message]
      properties:
        session_id:
          type: string
          format: uuid
          nullable: true
          description: Идентификатор диалога. Если не задан — сервер создаст новый.
          example: 2f6a9b3c-1d4e-4f8a-9c2b-5e7a1d0c3b66
        context:
          $ref: '#/components/schemas/GuideContext'
        message:
          type: string
          minLength: 1
          description: Сообщение пользователя (часто — одна из подсказок).
          example: Кому подарили это яйцо?
        history:
          type: array
          description: Опциональная история сообщений (если не используется session_id).
          items:
            $ref: '#/components/schemas/ChatMessage'
        language:
          type: string
          default: ru
          example: ru
        max_questions:
          type: integer
          minimum: 0
          maximum: 6
          default: 3
    ChatResponse:
      type: object
      required: [session_id, answer, suggested_questions]
      properties:
        session_id:
          type: string
          format: uuid
          example: 2f6a9b3c-1d4e-4f8a-9c2b-5e7a1d0c3b66
        answer:
          type: string
        suggested_questions:
          type: array
          items:
            type: string
        context:
          $ref: '#/components/schemas/GuideContext'

    # ── Озвучивание ──
...
    Event:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum:
            - app_open
            - hall_view
            - showcase_view
            - exhibit_view
            - recognition
            - chat_message
            - audio_play
            - suggestion_click
          example: exhibit_view
        exhibit_id:
          type: integer
          nullable: true
        hall_id:
          type: integer
          nullable: true
        label_slug:
          type: string
          nullable: true
        ts:
          type: string
          format: date-time
          nullable: true
        props:
          type: object
          nullable: true
          additionalProperties: true
    EventBatch:
      type: object
      required: [events]
      properties:
        session_id:
          type: string
          format: uuid
          nullable: true
        events:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Event'

    # ── Ошибки (формат FastAPI) ──
    Error:
      type: object
      description: Стандартная ошибка приложения (FastAPI `HTTPException`).
      required: [detail]
      properties:
        detail:
          type: string
          example: Экспонат не найден.
    ValidationErrorItem:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
          example: [body, message]
        msg:
          type: string
          example: field required
        type:
          type: string
          example: value_error.missing
    HTTPValidationError:
      type: object
      description: Ошибка валидации запроса в формате FastAPI.
      properties:
        detail:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorItem'